feat: add `PokemonRepository`

Adds a Pokémon repository and wires `Pokedex` to load Pokémon via `IPokemonRepository`.
Standardizes IDs to `uint32_t`, fixes ability/type data, returns `std::optional` for missing Pokémon,
and updates the UI to guard optional data.

- **New Features**
  - Introduced `IPokemonRepository` and in-memory `PokemonRepository` backed by `Pokemons` data; added `types::POISON`.
  - `Pokedex` now takes an `IPokemonRepository&` and returns `std::optional<Pokemon>`/`std::optional<PokemonDetails>`; `main` injects `PokemonRepository`.
  - Build: `pokemon_repository` exposes `pokemon` as PUBLIC and keeps `data` PRIVATE.
  - `UserInterface` now checks for missing Pokémon and empty types/evolutions before logging.

- **Migration**
  - Construct `Pokedex` with an `IPokemonRepository` and handle `std::optional` returns for Pokémon and details.
  - Update call sites to use `uint32_t` for Pokémon IDs, evolution IDs, and evolution levels.
  - Replace `abilities::OVERGROWTH` with `abilities::OVERGROW`.

+++ src/pokedex/model/evolution/Evolution.h
@@ -20,19 +20,19 @@ class Evolution
 {
 public:
     Evolution(
-        uint16_t from,
-        uint16_t to,
+        uint32_t  from,
+        uint32_t  to,
         const EvolutionMethod& method);
 
-    uint16_t from() const;
+    uint32_t  from() const;
     
-    uint16_t to() const;
+    uint32_t  to() const;
 
     const EvolutionMethod method() const;
 
 private:
-    uint16_t _from;
-    uint16_t _to;
+    uint32_t  _from;
+    uint32_t  _to;
     const EvolutionMethod _method;
 
 };
